Fix C-RET on list item bug
This function didn't work in some cases, e.g. right before a heading,
because `org-end-of-item` would take the point to the next line, where
`org-insert-item` no longer recognized a list. Substituting for
`end-of-line` fixes that.

However, this does mean that for situations like:

- A|
  - B

...in Doom it would create an item below B, whereas in my config it puts
one between A and B. However, I don't necessarily think that's a bad
thing, and might actually be more intuitive. It trips me up because I
got used to Doom's behaviour.