Bind M-x to SPC SPC
This is my third binding for M-x, the others being \ and x.
\ is not easily reachable on the Planck, and requires a modifier with
my current setup.
x is pretty convenient, being a single key, but I might want to use it
for other things.

Double SPC makes sense and is very easily reached. A double tap takes
only a little more effort than a single keystroke, but with SPC being
so big and so conveniently positioned, this is a non-issue.

I'll experiment more and decide which one(s) I want to keep.