Fix magit commit keybind
I use "c" for navigation, and I was assuming that "C" would be rebound
automatically through meow motion binds. That would normally be the
case, except magit already bound "C" to magit-commit-add-log, which
also commits, but adds extra information to the commit message, which
confused me for a while. Rebound "C" to just magit-commit.

Also rebound pulling to lowercase "f" for convenience.